fix(disclosure): address multi-agent review findings (iterations 1-15)
Contract:
- name and names are alternatives; both together is a recoverable caller
  error. Nulls are soft-absence at every level: null items become the
  same unresolvable empty entry as empty strings, never a whole-call
  type error.
- The length bound is enforced during the trim+dedup pass with
  borrowed-spelling dedup; the byte cap is enforced at parse time too —
  over-long spellings are stored only as their bounded 64-byte echo in a
  DescribeName::TooLong variant, so transient allocation is capped at
  batch-size x byte-cap and the too-long diagnostic survives.
- A bulk batch where every entry fails keeps the single-name verdict
  class (recoverable failure, TOOL_DESCRIBE_RETURNED_NO_SCHEMAS_MESSAGE),
  decided by the resolved-name set; batches with at least one schema
  succeed. Pinned end-to-end by an integration test scripting an
  all-fail batch.
- Disclosure bookkeeping is single-sourced in describe_one (selection
  log + resolved catalog name) and shared by both shapes, so it cannot
  diverge; bulk per-entry failures render through one error_entry shape;
  alias dedup returns None.
- Bridge names are rejected exactly once per shape (single-name
  pre-check ahead of the turn-state guard; bulk loop inline); the
  shared resolution helper carries only Unknown/Resolved — no
  production panic path.
- Single-name path is byte-exact pre-bulk on its model-visible surface,
  now with the exact "tool_describe returned schema" summary pinned.
- Bound constants are unambiguous (MAX_DESCRIBE_BATCH_SIZE vs per-name
  byte caps); advertised prose is deduplicated and pinned with exact
  bound phrases; every model-visible failure message is a constant.
  The PR body states the <=8-bulk vs >8-parallel trust asymmetry
  explicitly.

Structure:
- Module split follows the crate's port-module convention
  (tool_disclosure_port.rs + tool_disclosure_port/tests.rs);
  architecture-test allowlists and crate AGENTS.md keep their original
  paths; tests.rs carries the arch-exempt marker; pre-commit safety
  checks green.
- describe_resolution keeps the #5712 opacity boundary in one place;
  describe_output gives invoke_tool_describe one failure tail; the
  too-many message is derived from the bound constant.

Tests (17 unit + 32 integration, all green; clippy -D warnings clean;
289 architecture tests green): either/or rejection, null-as-absent at
both levels, exactly-max accept, post-dedup bound (nine copies of one
name collapse), all-error recoverable failure (unit + integration),
trim positive case, resolved alias dedup, denied alias per-spelling
entries, dotted-spelling disclosure on both shapes, empty/null/over-long
(multibyte echo, exact boundary) per-entry failures, exact single-name
summary pin, prose pins. The integration tier documents why
content-level schema delivery is pinned at unit tier: the model-visible
rendering strips describe payload keys and tool-result content lives in
the result store behind opaque refs.

Tracking: tool_disclosure_port.rs decomposition -> nearai/ironclaw#7383.
