+++ crates/loop/ironclaw_loop_host/prompts/tool_disclosure_protocol.md
@@ -5,7 +5,7 @@ When `tool_search` is present in your visible tool list, that list is a curated 
 In that case, when you need a capability and do not see a directly matching tool, do not assume it is unavailable and do not give up or tell the user you cannot do it. Discover the tool first:
 
 1. Call `tool_search` with a `query` describing what you need — a service name, an action, or a file type (for example `tool_search(query="github")` or `tool_search(query="send email")`). It returns matching tool names and one-line descriptions from the on-demand catalog.
-2. Load the schemas of every promising result in a **single response** — do not spend one model turn per tool. Prefer one bulk call, `tool_describe(names=["<tool-a>", "<tool-b>", "<tool-c>"])`, which returns each schema as its own entry; a name that is wrong or unavailable fails only its own entry, so the rest still come back. One call takes at most 8 names — for more candidates than that, emit several `tool_describe` calls in parallel in the same response rather than spreading them across turns. `tool_describe(name="<tool>")` remains valid for a single tool. Describing tools one per round-trip is the one thing to avoid.
+2. Load the schemas of every promising result in a **single response** — do not spend one model turn per tool. Prefer one bulk call, `tool_describe(names=["<tool-a>", "<tool-b>", "<tool-c>"])`, which returns each schema as its own entry; a name that is wrong or unavailable fails only its own entry, so the rest still come back, and two spellings of the same tool collapse to one entry. One call takes at most 8 names — for more candidates than that, emit several `tool_describe` calls in parallel in the same response rather than spreading them across turns. `tool_describe(name="<tool>")` remains valid for a single tool. Describing tools one per round-trip is the one thing to avoid.
 3. Invoke it with `tool_call(name="<tool>", arguments="{\"field\":\"value\"}")`, where `arguments` is a JSON object encoded as a string. Once you know a tool's exact name you may also call it directly by that name — approvals, policy, hooks, and safety run identically either way.
 
 When `tool_search` is present, always search before concluding a capability is unavailable. Only tell the user you cannot do something after `tool_search` returns nothing relevant.
